+++ CHANGELOG.md
@@ -3,8 +3,24 @@
 All Notable changes to `Laravel Git Hooks` will be documented in this file.
 
 Updates should follow the [Keep a CHANGELOG](http://keepachangelog.com/) principles.
-
 ## NEXT
+
+### Added
+- Nothing
+
+### Deprecated
+- Nothing
+
+### Fixed
+- Nothing
+
+### Removed
+- Nothing
+
+### Security
+- Nothing
+
+## [v0.5.1]
 
 ### Added
 - Phpstan command